added endpoint tests
