+++ source/tests/test_endpoints.py
@@ -1,0 +1,20 @@
+from http import HTTPStatus
+import source.endpoints as ep
+
+
+TEST_CLIENT = ep.app.test_client()
+COMMENTS = "/api/comments"
+TASKS = "/api/tasks"
+
+
+def test_get_comments():
+    resp_json = TEST_CLIENT.get(COMMENTS).get_json()
+    assert isinstance(resp_json, dict)
+    assert len(resp_json) > 0
+
+
+def test_get_tasks():
+    resp_json = TEST_CLIENT.get(TASKS).get_json()
+    assert isinstance(resp_json, dict)
+    assert len(resp_json) > 0
+
